requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# XAI libraries
shap>=0.42.0
lime>=0.2.0.1
eli5>=0.13.0

# Web frameworks
streamlit>=1.28.0
flask>=3.0.0
flask-cors>=4.0.0

# LLM Integration
openai>=1.0.0
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
scipy>=1.11.0
